+++ src/conf/config.rs
@@ -82,6 +82,12 @@ pub struct Config {
     /// Env vars to set for all job executions
     #[serde(default)]
     pub env: HashMap<String, String>,
+
+    /// Whether to beep when the job ends
+    pub beep_on_end: Option<bool>,
+
+    #[serde(default)]
+    pub sound: SoundConfig,
 }
 
 impl Config {