+++ src/internal.rs
@@ -1,5 +1,10 @@
 use {
-    crate::ScrollCommand,
+    crate::{
+        PlaySoundCommand,
+        ScrollCommand,
+        Volume,
+    },
+    lazy_regex::*,
     serde::{
         Deserialize,
         Deserializer,
@@ -37,6 +42,7 @@ pub enum Internal {
     Validate, // validate search entry
     NextMatch,
     PreviousMatch,
+    PlaySound(PlaySoundCommand),
 }
 
 impl Internal {
@@ -63,6 +69,7 @@ impl Internal {
             Self::Validate => "validate".to_string(),
             Self::NextMatch => "next match".to_string(),
             Self::PreviousMatch => "previous match".to_string(),
+            Self::PlaySound(_) => "play sound".to_string(),
         }
     }
 }
@@ -93,6 +100,13 @@ impl fmt::Display for Internal {
             Self::Validate => write!(f, "validate"),
             Self::NextMatch => write!(f, "next-match"),
             Self::PreviousMatch => write!(f, "previous-match"),
+            Self::PlaySound(PlaySoundCommand { name, volume }) => {
+                write!(f, "play-sound(")?;
+                if let Some(name) = name {
+                    write!(f, "name={},", name)?;
+                }
+                write!(f, "volume={})", volume)
+            }
         }
     }
 }
@@ -125,7 +139,27 @@ impl std::str::FromStr for Internal {
             "next-match" => Ok(Self::NextMatch),
             "previous-match" => Ok(Self::PreviousMatch),
             "copy-unstyled-output" => Ok(Self::CopyUnstyledOutput),
-            _ => Err("invalid internal"),
+            "play-sound" => Ok(Self::PlaySound(PlaySoundCommand::default())),
+            _ => {
+                if let Some((_, props)) = regex_captures!(r"^play[_-]sound\((.*)\)$", s) {
+                    let iter = regex_captures_iter!(r"([^=,]+)=([^=,]+)", props);
+                    let mut volume = Volume::default();
+                    let mut name = None;
+                    for (_, [prop_name, prop_value]) in iter.map(|c| c.extract()) {
+                        match prop_name {
+                            "name" => {
+                                name = Some(prop_value.to_string());
+                            }
+                            "volume" => {
+                                volume = prop_value.parse()?;
+                            }
+                            _ => return Err("invalid internal"),
+                        }
+                    }
+                    return Ok(Self::PlaySound(PlaySoundCommand { name, volume }));
+                }
+                Err("invalid internal")
+            }
         }
     }
 }
@@ -152,6 +186,7 @@ impl<'de> Deserialize<'de> for Internal {
 
 #[test]
 fn test_internal_string_round_trip() {
+    use crate::Volume;
     let internals = [
         Internal::Back,
         Internal::FocusSearch,
@@ -174,8 +209,22 @@ fn test_internal_string_round_trip() {
         Internal::Validate,
         Internal::NextMatch,
         Internal::PreviousMatch,
+        Internal::PlaySound(PlaySoundCommand::default()),
+        Internal::PlaySound(PlaySoundCommand {
+            name: None,
+            volume: Volume::new(50),
+        }),
+        Internal::PlaySound(PlaySoundCommand {
+            name: Some("beep-beep".to_string()),
+            volume: Volume::new(100),
+        }),
+        Internal::PlaySound(PlaySoundCommand {
+            name: None,
+            volume: Volume::new(0),
+        }),
     ];
     for internal in internals {
+        println!("testing {:?}", internal.to_string());
         assert_eq!(internal.to_string().parse(), Ok(internal));
     }
 }