+++ src/mission.rs
@@ -49,11 +49,6 @@ impl Mission<'_> {
         set
     }
 
-    /// the action bound to success on this job
-    pub fn on_success(&self) -> &Option<Action> {
-        &self.job.on_success
-    }
-
     pub fn is_success(
         &self,
         report: &Report,
@@ -234,6 +229,20 @@ impl Mission<'_> {
             .as_ref()
             .or(self.settings.ignored_lines.as_ref())
             .filter(|p| !p.is_empty())
+    }
+
+    pub fn sound_player_if_needed(&self) -> Option<SoundPlayer> {
+        if self.settings.sound.is_enabled() {
+            match SoundPlayer::new(self.settings.sound.get_base_volume()) {
+                Ok(sound_player) => Some(sound_player),
+                Err(e) => {
+                    warn!("Failed to initialise sound player: {e}");
+                    None
+                }
+            }
+        } else {
+            None
+        }
     }
 }